+++ src/applications/item-pile-inventory-app/item-pile-inventory-app.js
@@ -4,8 +4,10 @@ import * as Utilities from "../../helpers/utilities.js";
 import ItemPileConfig from "../item-pile-config/item-pile-config.js";
 import HOOKS from "../../constants/hooks.js";
 import * as Helpers from "../../helpers/helpers.js";
+import UserSelectDialog from "../dialogs/user-select-dialog/user-select-dialog.js";
+import SETTINGS from "../../constants/settings.js";
 
-export class ItemPileInventoryApp extends SvelteApplication {
+export default class ItemPileInventoryApp extends SvelteApplication {
 
   /**
    *
@@ -86,7 +88,7 @@ export class ItemPileInventoryApp extends SvelteApplication {
     if (canConfigure) {
       buttons = [
         {
-          label: "ITEM-PILES.Inspect.OpenSheet",
+          label: !Helpers.getSetting(SETTINGS.HIDE_ACTOR_HEADER_TEXT) ? "ITEM-PILES.Inspect.OpenSheet" : "",
           class: "item-piles-open-actor-sheet",
           icon: "fas fa-user",
           onclick: () => {
@@ -94,14 +96,16 @@ export class ItemPileInventoryApp extends SvelteApplication {
           }
         },
         {
-          label: "ITEM-PILES.ContextMenu.ShowToPlayers",
+          label: !Helpers.getSetting(SETTINGS.HIDE_ACTOR_HEADER_TEXT) ? "ITEM-PILES.ContextMenu.ShowToPlayers" : "",
           class: "item-piles-show-to-players",
           icon: "fas fa-eye",
-          onclick: () => {
-            const users = Array.from(game.users).filter(u => u.active && u !== game.user).map(u => u.id);
-            if (!users.length) {
+          onclick: async (event) => {
+            const activeUsers = Array.from(game.users).filter(u => u.active && u !== game.user).map(u => u.id);
+            if (!activeUsers.length) {
               return Helpers.custom_warning(game.i18n.localize("ITEM-PILES.Warnings.NoPlayersActive"), true);
             }
+            const users = event.altKey ? activeUsers : await UserSelectDialog.show({ excludeSelf: true });
+            if (!users || !users.length) return;
             Helpers.custom_notify(game.i18n.format("ITEM-PILES.Notifications.ShownToPlayers", { actor_name: this.actor.name }))
             return game.itempiles.API.renderItemPileInterface(this.actor, {
               userIds: users,
@@ -110,7 +114,7 @@ export class ItemPileInventoryApp extends SvelteApplication {
           }
         },
         {
-          label: "ITEM-PILES.HUD.Configure",
+          label: !Helpers.getSetting(SETTINGS.HIDE_ACTOR_HEADER_TEXT) ? "ITEM-PILES.HUD.Configure" : "",
           class: "item-piles-configure-pile",
           icon: "fas fa-box-open",
           onclick: () => {